code/requirements.txt
argparse
bs4
lxml
numpy
progressbar2
